Refactor progress tracker to Container/Component style
