+++ app/assets/javascripts/components/overview/my_articles/common/AssignedToLink.jsx
@@ -1,9 +1,6 @@
 import React from 'react';
 
-// Helper Components
-export const Separator = () => <span> •&nbsp;</span>;
-
-export const AssignedToLink = ({ name, members }) => {
+export default ({ name, members }) => {
   if (!members) return null;
 
   const label = <span key="label">{I18n.t(`assignments.${name}`)}: </span>;
@@ -19,12 +16,4 @@ export const AssignedToLink = ({ name, members }) => {
   });
 
   return [label].concat(links);
-};
-
-export const EditorLink = ({ editors }) => {
-  return <AssignedToLink members={editors} name="editors" />;
-};
-
-export const ReviewerLink = ({ reviewers }) => {
-  return <AssignedToLink members={reviewers} name="reviewers" />;
 };
