+++ app/assets/javascripts/components/overview/my_articles/common/ReviewerLink.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+// components
+import AssignedToLink from './AssignedToLink';
+
+export default ({ reviewers }) => {
+  return <AssignedToLink members={reviewers} name="reviewers" />;
+};
