+++ app/assets/javascripts/components/overview/my_articles/common/Separator.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default () => <span> •&nbsp;</span>;
