+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/Header/Actions/RemoveButton.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default ({ assignment, unassign }) => (
+  <div>
+    <button
+      onClick={() => unassign(assignment)}
+      className="button danger small"
+    >
+      Remove
+    </button>
+  </div>
+);
