+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/Header/Actions/index.jsx
@@ -1,0 +1,63 @@
+import React from 'react';
+
+// components
+import PageViews from './PageViews';
+import MarkAsIncompleteButton from './MarkAsIncompleteButton';
+import RemoveButton from './RemoveButton';
+
+import MainspaceChecklist from '../../../../../../../../common/mainspace_checklist.jsx';
+import FinalArticleChecklist from '../../../../../../../../common/final_article_checklist.jsx';
+import PeerReviewChecklist from '../../../../../../../../common/peer_review_checklist.jsx';
+import Feedback from '../../../../../../../../common/feedback.jsx';
+
+export default ({
+  article, assignment, courseSlug, current_user, isComplete, username,
+  isEnglishWikipedia, handleUpdateAssignment, refreshAssignments, unassign
+}) => {
+  const actions = [];
+
+  if (assignment.article_id) {
+    actions.push(<PageViews key="pageviews-button" article={article} />);
+  }
+
+  // Assigned article that does not yet exist in mainspace
+  if (isEnglishWikipedia()) {
+    const feedback = (
+      <Feedback
+        assignment={assignment}
+        current_user={current_user}
+        key="feedback-button"
+        username={username}
+      />
+    );
+    if (assignment.role === 0 && !assignment.article_id) {
+      actions.push(<MainspaceChecklist key="mainspace-button" />, feedback);
+    } else if (assignment.role === 0) {
+      actions.push(<FinalArticleChecklist key="final-article-button" />, feedback);
+    } else {
+      actions.push(<PeerReviewChecklist key="peer-review-button" />);
+    }
+  }
+
+  if (isComplete) {
+    return (
+      <section className="actions">
+        <PageViews key="pageviews-button" article={article} />
+        <MarkAsIncompleteButton
+          key="mark-incomplete-button"
+          assignment={assignment}
+          courseSlug={courseSlug}
+          handleUpdateAssignment={handleUpdateAssignment}
+          refreshAssignments={refreshAssignments}
+        />
+      </section>
+    );
+  }
+
+  return (
+    <section className="actions">
+      {actions}
+      <RemoveButton key="remove-button" assignment={assignment} unassign={unassign} />
+    </section>
+  );
+};
