+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/Header/Links/EditorLink.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import AssignedToLink from '../../../../../../common/AssignedToLink';
+
+export default ({ editors }) => {
+  return <AssignedToLink members={editors} name="editors" />;
+};
