+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/ProgressTracker/Step/Description.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default ({ content }) => (
+  <p className="step-description">{content}</p>
+);
