+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/ProgressTracker/Step/Reviewers.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+// components
+import ReviewerLink from '../../../../../../common/ReviewerLink';
+
+export default ({ reviewers }) => {
+  if (!reviewers) return null;
+
+  return (
+    <section className="step-members">
+      <ReviewerLink reviewers={reviewers} />
+    </section>
+  );
+};
