+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/ProgressTracker/Step/StepNumber.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default ({ index }) => (
+  <span className="step-number">{index + 1}</span>
+);
