+++ app/assets/javascripts/components/overview/my_articles/components/Categories/List/Assignment/ProgressTracker/Step/Title.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default ({ title }) => (
+  <h3 className="step-title">{title}</h3>
+);
