+++ app/assets/javascripts/components/overview/my_articles/components/Categories/index.jsx
@@ -1,0 +1,44 @@
+import React from 'react';
+import { IMPROVING_ARTICLE, NEW_ARTICLE, REVIEWING_ARTICLE } from '../../../../../constants/assignments';
+
+// components
+import List from './List';
+
+// Helper Functions
+const filterStatus = status => ({ article_status }) => {
+  return status === article_status;
+};
+
+const sortBy = key => (a, b) => {
+  return a[key] > b[key] ? 1 : -1;
+};
+
+// Main Component
+export default ({ assignments, course, current_user, wikidataLabels }) => {
+  const articles = {
+    new: assignments.filter(filterStatus(NEW_ARTICLE)).sort(sortBy('article_title')),
+    improving: assignments.filter(filterStatus(IMPROVING_ARTICLE)).sort(sortBy('article_title')),
+    reviewing: assignments.filter(filterStatus(REVIEWING_ARTICLE)).sort(sortBy('article_title'))
+  };
+
+  const listProps = { course, current_user, wikidataLabels };
+  return (
+    <>
+      {
+        articles.new.length
+          ? <List {...listProps} assignments={articles.new} title="Articles I will create" />
+          : null
+      }
+      {
+        articles.improving.length
+          ? <List {...listProps} assignments={articles.improving} title="Articles I'm updating" />
+          : null
+      }
+      {
+        articles.reviewing.length
+          ? <List {...listProps} assignments={articles.reviewing} title="Articles I'm peer reviewing" />
+          : null
+      }
+    </>
+  );
+};
