+++ app/assets/javascripts/components/overview/my_articles/components/NoAssignmentMessage.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export default () => (
+  <section className="no-assignment-message">
+    <p>You have not chosen an article to work on. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aenean molestie, lectus id consequat placerat, elit ligula sodales risus, non fringilla felis eros sed elit.</p>
+    <aside>
+      <a href="/training/students/finding-your-article" target="_blank" className="button ghost-button">
+        How to find an article
+      </a>
+      <a href="/training/students/evaluating-articles" target="_blank" className="button ghost-button">
+        Evaluating articles and sources
+      </a>
+    </aside>
+  </section>
+);
